Moved navigation components to their own directory
