+++ mobileApp/src/components/NavComponents/RootStack.tsx
@@ -2,13 +2,14 @@ import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import { getHeaderTitle } from "@react-navigation/elements";
 
 import BottomNav from "./BottomNav";
-import DetailsScreen from "./DetailsScreen";
-import { CustomAppBar } from "./CustomAppBar";
-import FormScreen from "./FormScreen";
-import FormAppBar from "../screens/Forms/FormAppBar";
+import DetailsScreen from "../DetailsScreen";
+import { CustomAppBar } from "../AppBars/CustomAppBar";
+import FormScreen from "../FormScreen";
+import FormAppBar from "../AppBars/FormAppBar";
 
 import { NativeStackHeaderProps } from "@react-navigation/native-stack";
-import { RootStackParamList } from "../NavigationTypes";
+import { RootStackParamList } from "../../NavigationTypes";
+import DetailsAppBar from "../AppBars/DetailsAppBar";
 
 const Stack = createNativeStackNavigator<RootStackParamList>();
 
@@ -32,7 +33,16 @@ function RootStack() {
                     ),
                 }}
             />
-            <Stack.Screen name="Details" component={DetailsScreen} />
+            <Stack.Screen
+                name="Details"
+                component={DetailsScreen}
+                options={{
+                    header: (props: NativeStackHeaderProps) => (
+                        <DetailsAppBar {...props} />
+                    ),
+                }}
+            />
+
             {/*TODO: Confirgure AppBar for DetailsScreen*/}
             <Stack.Screen
                 name="Forms"
